Suppress logger test messages
